+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added `pybricksdev oad info` command.
+- Added `pybricksdev oad flash` command.
 
 ## [1.0.0-alpha.50] - 2024-07-01